13/08/2022

+++ backend2/controllers/blogs.js
@@ -7,12 +7,18 @@ blogRouter.get('/', async(request, response) => {
     response.json(await Blog.find({}).populate('userId', {name: 1, username: 1}))
 })
 
-blogRouter.get('/:id', async(request, response) => response.json(await Blog.findById(request.params.id)))
-
-blogRouter.delete('/:id', async(request, response) => {
+blogRouter.get('/:id', async(request, response) => {
     if(!request.token) return response.status(401).json({error: 'Token missing or invalid'})
     const user = jwt.verify(request.token, process.env.SECRET)
-    const blog = await Blog.findById(request.params.id).populate('userId')
+    response.json(await Blog.findById(request.params.id))}
+)
+
+blogRouter.delete('/:id', async(request, response) => {
+    if(!request.token || request.token.length < 10) return response.status(401).json({error: 'Token missing or invalid'})
+    const user = jwt.verify(request.token, process.env.SECRET)
+    const blog = await Blog.findById(request.params.id).populate('userId', {name: 1, username: 1})
+    console.log(user)
+    console.log(blog)
     if(user){
         if(user.username == blog.userId.username){
             await Blog.findByIdAndRemove(request.params.id)
@@ -40,9 +46,10 @@ blogRouter.put('/:id', async(request, response) => {
 blogRouter.post('/', async(request, response) => {
     const body = request.body
     let token = request.get('Authorization')
-    if(!token) return response.status(400).json({error: 'Token missing or invalid'})
+    if(!token || token.length < 10) return response.status(401).json({error: 'Token missing or invalid'})
     token = token.substring(7)
     const obj = jwt.verify(token, process.env.SECRET)
+    console.log("Id: " + obj.id)
     const user = await User.findById(obj.id)
     if(!body.author && !body.title) return response.status(400).json({error: 'Author and title are REQUIRED fields'})
     if(!body.likes) body.likes = "0"
